Update prod task stats to new derivation model
Updates the task-stats derivations to use the new typescript module
definitions, so that we can re-apply them using the new
flowctl/control-plane. Also rips out the mock tasks, which were used to
ensure the appropriate ops collections got generated, with explicit
definitions of the ops collections now that this is allowed by
[this pr](https://github.com/estuary/flow/pull/516)
