--- ops-catalog/mock-tasks.flow.yaml
@@ -1,107 +1,0 @@
-# This file exists for two reasons:
-# - For prod.flow.yaml: to work around the lack of support for foreign entity resolution in the CLI
-# - For local.flow.yaml: because locally there would be no foreign entities to resolve anyway
-#
-# This includes a task for each tenant that has stats being materialized. The shards of all but one
-# of those are disabled in order to keep the data volume down, and since they don't actually need to
-# run in order to cause stats collections to get generated.
-collections:
-  acmeCo/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-  deepsync/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-  fenestra/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-  phil/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-  rockset/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-  wgd/greetings:
-    schema: greeting-schema.yaml
-    key: [/count]
-
-captures:
-  acmeCo/hello-world:
-    shards:
-      disable: false
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: acmeCo/greetings
-  deepsync/hello-world:
-    shards:
-      disable: true
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: deepsync/greetings
-  fenestra/hello-world:
-    shards:
-      disable: true
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: fenestra/greetings
-  phil/hello-world:
-    shards:
-      disable: true
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: phil/greetings
-  rockset/hello-world:
-    shards:
-      disable: true
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: rockset/greetings
-  wgd/hello-world:
-    shards:
-      disable: true
-    endpoint:
-      connector:
-        image: ghcr.io/estuary/source-hello-world:dev
-        config:
-          greetings: 5
-    bindings:
-      - resource:
-          stream: greetings
-          syncMode: incremental
-        target: wgd/greetings
-
